# monotickets-api/docs/contracts/OPENAPI_monotickets.yaml
openapi: 3.0.0
paths:
...
  /events/{id}:
    get:
      operationId: EventsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event retrieved successfully
        "404":
          description: Event not found
      security: *a3
      summary: Get event by ID
      tags: *a4
    patch:
      operationId: EventsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
      responses:
        "200":
          description: Event updated successfully
        "404":
          description: Event not found
      security: *a3
      summary: Update event
      tags: *a4
    delete:
      operationId: EventsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event deleted successfully
        "404":
          description: Event not found
      security: *a3
      summary: Delete event (soft delete)
      tags: *a4
...
  /events/{eventId}/invitations/generate:
    post:
      operationId: InvitationsController_generate
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInvitationsDto"
      responses:
        "201":
          description: Invitations generated successfully
          content:
            application/json:
              schema:
                example:
                  message: Successfully generated 50 invitations
                  count: 50
                  invitations: []
        "400":
          description: Invalid request or guests already have invitations
        "404":
          description: Event not found
      security: &a7
        - bearer: []
      summary: Generate invitations for guests
      tags: &a8
        - invitations
  /events/{eventId}/invitations:
    get:
      operationId: InvitationsController_findAll
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Invitations retrieved successfully
      security: *a7
      summary: Get all invitations for an event
      tags: *a8
  /events/{eventId}/invitations/stats:
    get:
      operationId: InvitationsController_getStats
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Invitation stats retrieved successfully
          content:
            application/json:
              schema:
                example:
                  totalInvitations: 150
                  sentInvitations: 120
                  deliveredInvitations: 100
                  pendingInvitations: 30
                  failedInvitations: 0
      security: *a7
      summary: Get invitation statistics for an event
      tags: *a8
  /invitations/{id}:
    get:
      operationId: InvitationsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Invitation retrieved successfully
        "404":
          description: Invitation not found
      security: *a7
      summary: Get invitation by ID
      tags: *a8
  /invitations/{id}/qr:
    get:
      operationId: InvitationsController_getQRToken
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: QR token generated successfully
          content:
            application/json:
              schema:
                example:
                  invitationId: 123e4567-e89b-12d3-a456-426614174000
                  qrToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  guestName: Juan Pérez
                  guestCount: 4
        "404":
          description: Invitation not found
      security: *a7
      summary: Get JWT-based QR token for invitation
      tags: *a8
  /invitations/qr/{qrToken}:
    get:
      operationId: InvitationsController_findByQRToken
      parameters:
        - name: qrToken
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Invitation retrieved successfully
        "404":
          description: Invitation not found
      security: *a7
      summary: Get invitation by QR token (legacy)
      tags: *a8
  /invitations/send:
    post:
      operationId: InvitationsController_send
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInvitationsDto"
      responses:
        "200":
          description: Invitations sent successfully
          content:
            application/json:
              schema:
                example:
                  message: Successfully sent 50 invitations
                  count: 50
                  method: SMS
                  invitations: []
        "400":
          description: Invalid request
      security: *a7
      summary: Send invitations via SMS/Email/WhatsApp
      tags: *a8
...
  /events/{eventId}/rsvp-config:
    get:
      operationId: RsvpController_getConfig
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: RSVP config retrieved successfully
          content:
            application/json:
              schema:
                example:
                  eventId: 550e8400-e29b-41d4-a716-446655440000
                  allowRsvp: true
                  rsvpDeadlineDays: 0
                  revocationWindowDays: 20
                  updatedAt: 2025-12-13T21:00:00.000Z
        "404":
          description: Event not found
      security: &a14
        - bearer: []
      summary: Get RSVP configuration for an event
      tags: &a15
        - rsvp
    post:
      operationId: RsvpController_createConfig
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRsvpConfigDto"
      responses:
        "201":
          description: RSVP config created successfully
          content:
            application/json:
              schema:
                example:
                  eventId: 550e8400-e29b-41d4-a716-446655440000
                  allowRsvp: true
                  rsvpDeadlineDays: 5
                  revocationWindowDays: 20
        "404":
          description: Event not found
      security: *a14
      summary: Create/Update RSVP configuration
      tags: *a15
  /events/{eventId}/confirmations:
    get:
      operationId: RsvpController_getConfirmations
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Confirmations retrieved successfully
          content:
            application/json:
              schema:
                example:
                  eventId: 550e8400-e29b-41d4-a716-446655440000
                  total: 150
                  confirmed: 120
                  declined: 10
                  pending: 20
                  confirmationRate: 80
                  guests: []
        "404":
          description: Event not found
      security: *a14
      summary: Get RSVP confirmations list
      tags: *a15
...
info:
  title: Monotickets Platinum API
  description: API oficial para Monotickets Platinum (invited events, RSVPs,
    delivery y scanner).
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateEventDto:
      type: object
      properties:
        name:
          type: string
          description: Event name
          example: Boda de Juan y María
        date:
          type: string
          description: Event date
          example: 2025-12-25T18:00:00.000Z
        location:
          type: string
          description: Event location
          example: Jardín Las Rosas, CDMX
        description:
          type: string
          description: Event description
          example: Celebración de nuestra boda
        plannerId:
          type: string
          description: Planner user ID
          example: 550e8400-e29b-41d4-a716-446655440000
...
    GenerateInvitationsDto:
      type: object
      properties:
        guestIds:
          description: Array of guest IDs to generate invitations for
          example:
            - 550e8400-e29b-41d4-a716-446655440000
          type: array
          items:
            type: string
      required:
        - guestIds
    SendInvitationsDto:
      type: object
      properties:
        invitationIds:
          description: Array of invitation IDs to send
          example:
            - 550e8400-e29b-41d4-a716-446655440000
          type: array
          items:
            type: string
        method:
          type: string
          description: Delivery method
          enum:
            - SMS
            - EMAIL
            - WHATSAPP
          example: SMS
      required:
        - invitationIds
...
    UpdateRsvpConfigDto:
      type: object
      properties:
        allowRsvp:
          type: boolean
          description: Allow RSVP confirmations
        rsvpDeadlineDays:
          type: number
          description: Days before event to close RSVP (0 = Event Day)
        revocationWindowDays:
          type: number
          description: Days after receipt to allow revocation